Rename PropagationFormatterInterface -> PropagatorInterface
Also rename Formatter -> Propagator.
