+++ src/Trace/RequestHandler.php
@@ -24,7 +24,7 @@ use OpenCensus\Trace\Tracer\ContextTracer;
 use OpenCensus\Trace\Tracer\ExtensionTracer;
 use OpenCensus\Trace\Tracer\NullTracer;
 use OpenCensus\Trace\Tracer\TracerInterface;
-use OpenCensus\Trace\Propagation\PropagationFormatterInterface;
+use OpenCensus\Trace\Propagator\PropagatorInterface;
 
 /**
  * This class manages the logic for sampling and reporting a trace within a
@@ -77,7 +77,7 @@ class RequestHandler
      *
      * @param ReporterInterface $reporter How to report the trace at the end of the request
      * @param SamplerInterface $sampler Which sampler to use for sampling requests
-     * @param PropagationFormatterInterface $propagator TraceContext propagator
+     * @param PropagatorInterface $propagator TraceContext propagator
      * @param array $options [optional] {
      *      Configuration options. See
      *      {@see OpenCensus\Trace\TraceSpan::__construct()} for the other available options.
@@ -88,7 +88,7 @@ class RequestHandler
     public function __construct(
         ReporterInterface $reporter,
         SamplerInterface $sampler,
-        PropagationFormatterInterface $propagator,
+        PropagatorInterface $propagator,
         array $options = []
     ) {
         $this->reporter = $reporter;
