+++ tests/unit/Trace/RequestHandlerTest.php
@@ -22,7 +22,7 @@ use OpenCensus\Trace\RequestHandler;
 use OpenCensus\Trace\Reporter\ReporterInterface;
 use OpenCensus\Trace\Sampler\SamplerInterface;
 use OpenCensus\Trace\Tracer\NullTracer;
-use OpenCensus\Trace\Propagation\HttpHeaderFormatter;
+use OpenCensus\Trace\Propagator\HttpHeaderPropagator;
 
 /**
  * @group trace
@@ -49,7 +49,7 @@ class RequestHandlerTest extends \PHPUnit_Framework_TestCase
         $rt = new RequestHandler(
             $this->reporter->reveal(),
             $this->sampler->reveal(),
-            new HttpHeaderFormatter()
+            new HttpHeaderPropagator()
         );
         $rt->inSpan(['name' => 'inner'], function () {});
         $rt->onExit();
@@ -71,7 +71,7 @@ class RequestHandlerTest extends \PHPUnit_Framework_TestCase
         $rt = new RequestHandler(
             $this->reporter->reveal(),
             $this->sampler->reveal(),
-            new HttpHeaderFormatter(),
+            new HttpHeaderPropagator(),
             [
                 'headers' => [
                     'REQUEST_URI' => '/some/uri',
@@ -111,7 +111,7 @@ class RequestHandlerTest extends \PHPUnit_Framework_TestCase
         $rt = new RequestHandler(
             $this->reporter->reveal(),
             $this->sampler->reveal(),
-            new HttpHeaderFormatter(),
+            new HttpHeaderPropagator(),
             [
                 'headers' => [
                     'HTTP_X_CLOUD_TRACE_CONTEXT' => '12345678901234567890123456789012/5555;o=1'
@@ -129,7 +129,7 @@ class RequestHandlerTest extends \PHPUnit_Framework_TestCase
         $rt = new RequestHandler(
             $this->reporter->reveal(),
             $this->sampler->reveal(),
-            new HttpHeaderFormatter(),
+            new HttpHeaderPropagator(),
             [
                 'headers' => [
                     'HTTP_X_CLOUD_TRACE_CONTEXT' => '12345678901234567890123456789012;o=1'
@@ -146,7 +146,7 @@ class RequestHandlerTest extends \PHPUnit_Framework_TestCase
         $rt = new RequestHandler(
             $this->reporter->reveal(),
             $this->sampler->reveal(),
-            new HttpHeaderFormatter(),
+            new HttpHeaderPropagator(),
             [
                 'headers' => [
                     'HTTP_X_CLOUD_TRACE_CONTEXT' => '12345678901234567890123456789012;o=0'
